+++ storage/blockdevice/source/ExhaustibleBlockDevice.cpp
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "ExhaustibleBlockDevice.h"
+#include "blockdevice/ExhaustibleBlockDevice.h"
 #include "platform/mbed_atomic.h"
 #include "platform/mbed_assert.h"
 
@@ -209,4 +209,3 @@ const char *ExhaustibleBlockDevice::get_type() const
 }
 
 } // namespace mbed
-
