+++ storage/blockdevice/source/ProfilingBlockDevice.cpp
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "ProfilingBlockDevice.h"
+#include "blockdevice/ProfilingBlockDevice.h"
 #include "stddef.h"
 
 namespace mbed {
@@ -131,4 +131,3 @@ const char *ProfilingBlockDevice::get_type() const
 }
 
 } // namespace mbed
-
